thermal: odroid-xu4: ignore invalid temperatures from broken TMU
The temperatue value is invalid when the TMU is broken from electronic
damage. In that case, XU4 shuts down due to reaching the critical
temperature even if the real temperature is not high.

This change is to avoid such cases by checking whether
the temperature value is valid or not. If the temperature value is
invalid, then the value is set to a constant.

Signed-off-by: Brian Kim <brian.kim@hardkernel.com>
Signed-off-by: memeka <mihailescu2m@gmail.com>